refactor(ui): extract cookies effect business logic into services/cookies (ADR 0004 §4)
Move jar-picking heuristics, Set-Cookie parsing/capture, and
encrypted-file persistence out of store/effects/cookies.ts into three
focused modules under services/cookies/:

- jar-picker.ts: pure `pickOwningJar` — resolves primary variable-set
  jar from project + preference state, no side effects.
- capture.ts: `captureSetCookies` + `extractCookiesFromResponse` —
  parse Set-Cookie headers from completed flight responses.
- persistence.ts: `loadAndHydrateCookies` + `persistCookieJars` +
  `pruneAndCap` — sealed-envelope file I/O via ipcEncryptionService /
  ipcFsService.
- index.ts: explicit named re-exports, no wildcard barrels.

The effect file shrinks from 279 LOC to 58 LOC (3 listeners, each a
single delegating call). No action types, action creator names, or
observable behaviour changed.
